+++ CHANGELOG.md
@@ -4,6 +4,7 @@ BUG FIXES:
 
 * resource/aws_autoscaling_attachment: Retry errors like `ValidationError: Trying to update too many Load Balancers/Target Groups at once. The limit is 10` when creating or deleting resource ([#26654](https://github.com/hashicorp/terraform-provider-aws/issues/26654))
 * resource/aws_instance: Prevents errors in ISO regions when not using DisableApiStop attribute ([#26745](https://github.com/hashicorp/terraform-provider-aws/issues/26745))
+* resource/aws_ssoadmin_managed_policy_attachment: Prevent existing managed policy on permission set from being attached again causing InternalFailure due to infinite loop. ([#26138](https://github.com/hashicorp/terraform-provider-aws/issues/26138))
 
 ## 4.30.0 (September  9, 2022)